add artist page for youtube data source
Artists page now exists for the YouTube data source. It's a bit useless as it can't display videos due to what may be a bug in the node module (or maybe is just me being stupid, we shall find out soon). Anyways, the YouTube data source is getting more and more fleshed out!